### 2020-11-23 #96 @whedon
[PRE REVIEW]: An Open-Source Active Learning Curriculum for Data Science
**Submitting author:** @zdelrosario (<a href="http://orcid.org/0000-0003-4676-1692">Zachary del Rosario</a>)
**Repository:** <a href="https://github.com/zdelrosario/data-science-curriculum" target ="_blank">https://github.com/zdelrosario/data-science-curriculum</a>
**Version:** v0.1

@zdelrosario if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
