### 2020-11-23 #96 @labarba
hi @zdelrosario 👋  — it looks like you missed that you need to write a paper to describe your submission. See our documentation for [what your paper should contain](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain).
