### 2020-11-23 #96 @zdelrosario
@labarba thanks for the heads up! I'd written the paper to my development branch but forgot to merge that in. Trying again....
