### 2020-11-29 9afe0f7
add comment on formula notation
