+++ exercises/e-model01-intro-master.Rmd
@@ -163,6 +163,7 @@ fit_q1 %>% tidy()
 - What kind of variable is `carat`? What is `cut`?
 - Can we "multiply" `cut` by a constant? Why or why not?
 - Describe the difference between the `term`s for `carat` and `cut`.
+- How many `cut` terms does the model have? How many levels does `cut` have?
 <!-- task-end -->
 <!-- solution-begin -->
 - `carat` is a continuous variable, while `cut` is an (ordinal) factor; it only takes fixed non-numerical values.
