+++ exercises/e-vis06-themes-master.Rmd
@@ -13,6 +13,7 @@ reading: 0
 
 ```{r setup, include=FALSE}
 # knitr options
+library(tidyverse)
 knitr::opts_chunk$set(echo = TRUE)
 ```
 
