### 2021-05-26 #117 @whedon
[REVIEW]: An Open-Source Active Learning Curriculum for Data Science
**Submitting author:** @zdelrosario (<a href="http://orcid.org/0000-0003-4676-1692">Zachary del Rosario</a>)
**Repository:** <a href="https://github.com/zdelrosario/data-science-curriculum" target ="_blank">https://github.com/zdelrosario/data-science-curriculum</a>
**Version:** v0.1

- [x] **Version:** Does the release version given match the repository release (v0.1)?

---

- [x] **Version:** Does the release version given match the repository release (v0.1)?
