### 2021-05-26 95c2516
add missing DOIs
