+++ paper.bib
@@ -42,7 +42,8 @@
   volume={30},
   number={9},
   pages={1--18},
-  year={2013}
+  year={2013},
+  doi={10.18260/1-2--22585}
 }
 
 @book{lang2016small,
@@ -98,6 +99,7 @@
   number = {4},
   pages = {1--34},
   url = {http://www.jstatsoft.org/v64/i04/},
+  doi = {10.18637/jss.v064.i04}
 }
 
 @Manual{gapminder,
