### 2021-05-26 #117 @zdelrosario
(Update DOI for "The Flipped Classroom" and fitdistrplus....)
