### 2021-06-17 zdelrosario/data-science-curriculum#24 @pschloss
License
For the content, the JOSE review checklist indicates that a Creative Commons license should be used. MIT is more appropriate for code
