### 2021-06-17 zdelrosario/data-science-curriculum#25 @pschloss
Statement of Need
The statement of need could be more clear in the README and in the paper. There are active learning materials out there - e.g., The Carpentries, Codeacademy. What is the particular niche that your materials are filling?

I'm not sure that the "Notably, if teachers held ..." sentence is constructive. Surely the published books, blog posts, and videos have instructional value even if they aren't examples of active learning materials?
