### 2021-06-17 zdelrosario/data-science-curriculum#27 @pschloss
Usage
I think learners could use more support to get going. I wasn't sure whether I'm supposed to be working within the Rmd files or if make should have generated html files that I worked with. When I didn't find html files, I worried that something was wrong with the Makefile.

The links in the README and the files in the paper refer to the raw versions of the files in `exercises/` rather than `exercises_sequenced/`. I think it would be better, from the learner's perspective to refer to things in `exercises_sequenced/` since that's the directory they are supposed to be working in.

Perhaps the repository could include a Rproj file so that learners can open Rstudio in the correct working directory?
