### 2021-06-17 #117 @pschloss
I completed my first review and posted issues on the [target repository](https://github.com/zdelrosario/data-science-curriculum). I like the overall format and goals of the materials and think they represent a unique contribution. The issues I filed concerned...

* The [License](https://github.com/zdelrosario/data-science-curriculum/issues/24)
* The [Statement of Need](https://github.com/zdelrosario/data-science-curriculum/issues/25)
* The [Installation Instructions](https://github.com/zdelrosario/data-science-curriculum/issues/26)
* [Usage](https://github.com/zdelrosario/data-science-curriculum/issues/27)

I look forward to reviewing the revised materials!
