### 2021-07-02 zdelrosario/data-science-curriculum#27 @firasm
Adding my comments to this issue because it's related to the Usage:

- [ ] Step 5 of ["How to Use This Repo"](https://github.com/zdelrosario/data-science-curriculum#how-to-use-this-repo) should make it clear the directory name is `exercises` not `assignments`.
	- [ ] I don't quite understand the 'dXY' prefix, is this from a previous version? After I run make, I get file names like `e-stat03-descriptive-master.Rmd`
	- [ ] I agree with @pschloss, could you remove the requirement for (novice) users to run `make` ? If you're worried about the exercises being out-of-date from the source, you could consider running a GitHub Action that automates running (and committing) of the exercises after every commit.
