### 2021-07-02 zdelrosario/data-science-curriculum#25 @firasm
Adding my comments to this issue so the reviewer feedback is consolidated in one place:

## Statement of need

- [ ] Nice, if a bit concise. I would suggest though that if you're going to name resources, it would be helpful to a bit more selective. It's not so useful to point out "Medium Blog posts" (but statquest is a good specific reference). Others might be data8, learnr project, the "hadley books" (https://r4ds.had.co.nz), data science: a first introduction (https://ubc-dsci.github.io/introduction-to-datascience/). Not necessarily advocating including these, these are just some that I know.

- [ ] Glad to see the Freeman reference advocating for active learning in the classroom. Might be worthwhile to make the point that good resources are essential for self-study outside of lecture time, but in the classroom active learning shows substantial learning benefits.

- [ ] May be worth discussing the importance of open education resources (OER) to making learning more equitable.
