### 2021-07-02 zdelrosario/data-science-curriculum#28 @firasm
Cut new release after feedback is implemented?
Perhaps when all the feedback is implemented, it might be worth creating a new release ?

- [ ] There have been some commits since Oct 21, 2020, might be worth cutting a new release
