### 2021-07-02 zdelrosario/data-science-curriculum#29 @firasm
Pedagogical Design
From the paper,

"The full set of desired learning outcomes is documented in the project repository"

- [ ] Where are the learning outcomes in the repository? Consider a direct link into the paper?

- [ ] Add a citation for flipped classroom
    - Perhaps this one? https://peer.asee.org/the-flipped-classroom-a-survey-of-the-research; any one will do.
