### 2021-07-02 zdelrosario/data-science-curriculum#30 @firasm
Overall Feedback
Here are my overall comments with some suggestions. I've put it all in one issue, because some are overlapping.

## Overall

- [ ] I really appreciate that all your material is open under an MIT license! Thank you for doing this!

- [ ] Though I did not go through every exercise, the quality of your material is quite high overall.

- [ ] Having things in a GitHub repo is nice, it might make your content more accessible and discoverable if it was set up as a Bookdown (https://bookdown.org/yihui/bookdown/) or Jupyter Book (https://jupyterbook.org/intro.html)

- [ ] I really think this work would benefit from a better description of what it is you're offering to the community. Frankly, `paper.md` sells the content in your repository well, well short!

	- [ ] This is related to the "Does the paper describe the learning materials and sequence?" point in the checklist

	- [ ] "Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?" --> this is also missing from the paper

	- [ ] "Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?" --> Also missing from the paper

- [ ] You could provide a schematic or a sketch of the key "modules" (communication, data, model, reproducibility, setup, statistics, visualizations) and a suggested ordering of going through them. You could also bring in some pedagogical reasons (mentioned in your repo) for your suggestions such as blocked practice.

- [ ] If you have the list of learning outcomes already prepared and ready, could you consolidate them into a single file so instructors could get a sense of what your material is all about without opening every file?
