### 2021-07-02 zdelrosario/data-science-curriculum#31 @firasm
Sequence
## Sequence.md

- [ ] It's nice that you have the sequence, but it would be better for learners if the list of files were linked. (This comment should be ignored if you choose to create a bookdown)
