### 2021-07-02 #117 @firasm
Hello @zdelrosario!

Thanks for creating such a great resource! In some cases I've added my comments to existing issues so the reviewer feedback is consolidated in one place:

Links to a few issues:

- [Statement of need](https://github.com/zdelrosario/data-science-curriculum/issues/25)
- [Installation Instructions](https://github.com/zdelrosario/data-science-curriculum/issues/26)
- [Usage](https://github.com/zdelrosario/data-science-curriculum/issues/27)
- [New Release](https://github.com/zdelrosario/data-science-curriculum/issues/28)
- [Pedagogical Design](https://github.com/zdelrosario/data-science-curriculum/issues/29)
- [Sequence](https://github.com/zdelrosario/data-science-curriculum/issues/31)
- [Overall Feedback](https://github.com/zdelrosario/data-science-curriculum/issues/30)

Let me know when there's something new you'd like me to look at, or if you need help implementing any of my suggestions. If any of the suggestions seem particularly onerous or not rewarding, let's discuss!
