### 2021-07-07 #117 @zdelrosario
Hi @pschloss and @firasm ! Many thanks for the detailed feedback; I am looking forward to addressing these comments, as I believe they will *substantially* improve the work. Note that I have a few clarifying questions on some of your comments; I'll direct those to the relevant issues.
