### 2021-07-07 zdelrosario/data-science-curriculum#26 @zdelrosario
I agree that this is a weak point of the materials. I am very interested in using a GitHub Action to automate the build of the exercises from source; however, I've tried using a similar implementation on a separate repo in the past, and have found it difficult to avoid introducing irritating merge conflicts. @firasm , can you point me towards any example workflows (like what you're describing) that build, commit, and push files with a GitHub Action?
