### 2021-07-07 zdelrosario/data-science-curriculum#26 @firasm
Certainly @zdelrosario, [here's an example](https://github.com/open-resources/physics_bank/blob/main/.github/workflows/deploy.yml) that builds a [`jupyterbook`](http://jupyterbook.org) and then pushes the HTML to my server (handled using GitHub Organization Secrets). [Here's the result of the pushes](https://firas.moosvi.com/oer/physicsbank/files/home.html).

I think the trick to avoid merge conflicts is to clone the repo with the GH action. Here's the [action I rely on](https://github.com/EndBug/add-and-commit).

Another option is to have two repos, one with the source (instructor version), and one with the build materials (public/user) version. If you go that route, [this action is useful](https://github.com/dmnemec/copy_file_to_another_repo_action). This takes files from an repo, clones another repo locally (in the GH worker), adds the new files, and pushes it up.

Good luck!
