### 2021-07-12 3551913
start EDA vs presentation quality exercise
