### 2021-08-07 be85724
add control chart exercise
