+++ exercises/e-vis10-control-chart-master.Rmd
@@ -1,0 +1,342 @@
+---
+title: "Visualization: Control Charts"
+author: Zach del Rosario
+date: 2020-08-06
+output: github_document
+time: 30
+reading: 0
+---
+
+*Purpose*: Remember in c02-michelson (q4) when you studied a *control chart*? Now that we've learned about confidence intervals, we can more formally study control charts. These are a key tool in [statistical process control](https://en.wikipedia.org/wiki/Statistical_process_control), which is how manufacturers rigorously track and control the quality of manufactured goods. Control charts help a process manager track when something has gone wrong in a manufacturing line, and are used to determine when a process is running smoothly.
+
+*Reading*: [Example use](https://www.itl.nist.gov/div898/handbook/mpc/section3/mpc3521.htm) of a control chart, based on NIST mass calibration data. (Optional)
+
+*Prerequisites*: c02-michelson, e-stat06-clt
+
+```{r setup}
+library(tidyverse)
+```
+
+# Control Chart: Example
+
+Below is an example of a control chart: A control chart is used to help detect when something out-of-the-ordinary occurred. Essentially, it is a tool to help us determine when something non-random happened so we can plan a follow-up study and prevent that non-random event from happening in the future. To do that detection work, we look for patterns.
+
+Note that two kinds of patterns have been highlighted below: an outlier that lies outside the *control limits*, and a "run" of batch means that all lie off-center (to one side of the "grand mean," the solid line).
+
+![control chart](./images/control-chart.png)
+
+## Control chart steps
+
+To construct and use a control chart, follow these steps:
+
+1. Group individual observations into consecutive batches, say 4 to 10 observations. If parts are manufactured in batches, then use those groups.
+2. Take the mean of each batch, compute the "grand mean" based on all of the data, and establish "control limits" based on a confidence interval for the batch means (where $n$ is your batch size).
+3. Plot each batch mean sequentially, and visually indicate the control limits and grand mean on your plot.
+4. Compare each batch mean against the control limits and grand mean. Look for patterns to suggest batches where something out-of-the-ordinary happened. Some examples include:
+- Batches that fall outside the control limits
+- Consecutive batches that lie above or below the mean
+- Persistent up-and-down patterns
+
+If there are no coherent patterns and if only an expected number of batch means fall outside the control limits, then there is no evidence for non-random behavior. A process free of any obvious non-random behavior is said to be under *statistical control*, or to be a  [stable process](https://en.wikipedia.org/wiki/Statistical_process_control#Stable_process).
+
+If you *do* detect something out-of-the-ordinary, use the batch index to go investigate those cases in greater detail. **A control chart helps you *detect* when something went wrong---it does *not* tell you what went wrong.**
+
+Like any form of EDA, it is also a good idea to experiment with different batch sizes.
+
+## Data Preparation
+
+To illustrate the control chart concept, let's first generate some data that is completely random.
+
+```{r generate-data}
+set.seed(101)
+df_data <- 
+  tibble(X = rnorm(n = 1000))
+```
+
+Following Step 1, we need to assign batch identifiers to group the data.
+
[…]
