+++ exercises/e-model06-source-master.Rmd
@@ -1,0 +1,156 @@
+---
+title: "Variability Quadrants: Source"
+author: Zachary del Rosario
+date: 2021-08-06
+output: github_document
+time: 30
+reading: 0
+---
+
+*Purpose*: 
+
+*Prerequisites*: (None)
+
+```{r setup, include=FALSE}
+# knitr options
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+```{r library}
+## Note: No need to edit this chunk!
+library(tidyverse)
+```
+
+# Variability
+
+## The Cause-Source Quadrants
+
+![Variability quadrants](./images/cause-source-quadrants.png)
+
+In this exercise, we'll focus on the vertical axis: the **source** of variability.
+
+## Real vs Induced Source
+
+![Variability quadrants](./images/noise-v-deviation.png)
+
+# Manufacturing structural steel components
+
+*The Context*: An manufacturer is producing cast-steel parts for a landing gear. The part in question takes a heavy load, and if it fails it will disable the aircraft on the ground. These parts will be manufactured in bulk; approximately 500 will be made and installed in commercial aircraft that will operate for decades.
+
+*The Target*: The strength of each steel component---as-manufactured---will ultimately determine whether each aircraft is safe.
+
+## Scenarios
+
+### __q1__ Imagine the manufacturer selects one part and performs multiple non-destructive tensile tests on that single part, under similar conditions. The measured elasticity from each test is slightly different. Is this variability real or induced?
+
+<!-- task-begin -->
+- real or induced?
+- (Why?)
+<!-- task-end -->
+<!-- solution-begin -->
+- Induced
+- The properties of the component are essentially set at maunfacturing time; if multiple measurements on the same part return different values, then the variability is most likely induced by the measurement process.
+<!-- solution-end -->
+
+### __q2__ Imagine the manufacturer selects multiple parts and---for each part---performs multiple non-destructive tensile tests, all under similar conditions. The measured elasticity values for each part are averaged to provide a more reliable estimate for each part. Upon comparing the parts, each averaged value is fairly different. Is this variability real or induced?
+
+<!-- task-begin -->
+- real or induced?
+- (Why?)
+<!-- task-end -->
+<!-- solution-begin -->
+- Real
+- The properties of the component are essentially set at maunfacturing time; but no manufacturing process can create items with identical properties. Particularly if variability remains after induced variability has been controlled and eliminated (as described in the prompt), then the remaining variability is real.
+<!-- solution-end -->
+
+### __q3__ Now the manufacturer selects multiple parts and performs a destructive tensile test to characterize the strength of each part, with tests carried out under similar conditions. The measured strength values exhibit a fair amount of variability. Is this variability real or induced?
+
+<!-- task-begin -->
+- real or induced?
+- (Why?)
+<!-- task-end -->
+<!-- solution-begin -->
+- Without more information, it is impossible to say. It is likely a combination of real and induced sources.
+- Real variability can arise from the manufacturing process, and induced variability can arise from the measurement. Since the measurement is destructive, we cannot use multiple measurements to control the induced variability.
+- Note that it would generally be conservative to treat all of the variability in a strength as real; this would lead to parts that are heavier but safer than they need to be.
+<!-- solution-end -->
+
+## Analyzing Data
+
+The following generates data with both *noise* and *deviation*
+
+```{r gen-data}
+set.seed(101)
+df_meas <- 
+  map_dfr(
+    1:30,
+    function(i) {
[…]
