+++ exercises/e-model06-source-master.Rmd
@@ -7,9 +7,9 @@ time: 30
 reading: 0
 ---
 
-*Purpose*: 
+*Purpose*: All real data have variability: repeated measurements of "the same" quantity tend to result in different values. To help you recognize different kinds of variability and choose a reasonable analysis procedure based on the kind of variability, you will learn about different *sources* of variability in this exercise.
 
-*Prerequisites*: (None)
+*Prerequisites*: e-stat03-descriptive, c08-structures
 
 ```{r setup, include=FALSE}
 # knitr options
@@ -23,21 +23,34 @@ library(tidyverse)
 
 # Variability
 
+As we've seen in this course, real data exhibit *variability*; that is, repeated measurements of "the same" quantity that result in different values. Variability can arise due to a variety of reasons, and different kinds of variability should be analyzed in different ways. To help make this determination, we're going to study a theoretical framework for variability.
+
 ## The Cause-Source Quadrants
+
+The *cause-source quadrants* organize variability into four distinct categories. Today, we're going to focus on the *source* axis, and limit our attention to *chance causes*.
 
 ![Variability quadrants](./images/cause-source-quadrants.png)
 
-In this exercise, we'll focus on the vertical axis: the **source** of variability.
+- *Cause* is an idea from statistical quality control; a *chance cause* is understood to be random, while an *assignable cause* is thought to be traceable and preventable. We will explore cause further in the next exercise.
+- *Source* is an idea from statistical education theory; this concept is explained further below.
 
 ## Real vs Induced Source
 
+The idea of *source* can only be understood in the distinction between a *target value* and a *measured value*: The *target value* is the quantity that we are seeking to study, while the *measured value* is a possibly-corrupted version of our target. The key insight is that **variability can occur both in the target value, and in the measured value**.
+
 ![Variability quadrants](./images/noise-v-deviation.png)
+
+As a simple example: based on our current understanding of physics, the speed of light `c` is a [constant value](https://en.wikipedia.org/wiki/Speed_of_light). Therefore, any variability we see in measurements of `c` are understood to be *induced variability*; real variability in `c` is not considered to be possible.
+
+Conversely, our current understanding of physics is that quantum phenomena are [fundamentally unpredictable](https://en.wikipedia.org/wiki/Quantum_mechanics), and can only be described in a statistical sense. This means that quantum phenomena exhibit real variability.
+
+Other physical quantities exhibit both real and induced variability. Since the concept of *source* relies on a target value, the only way we can make progress with this concept is to consider a specific scenario in detail.
 
 # Manufacturing structural steel components
 
 *The Context*: An manufacturer is producing cast-steel parts for a landing gear. The part in question takes a heavy load, and if it fails it will disable the aircraft on the ground. These parts will be manufactured in bulk; approximately 500 will be made and installed in commercial aircraft that will operate for decades.
 
-*The Target*: The strength of each steel component---as-manufactured---will ultimately determine whether each aircraft is safe.
+*The Target*: The strength of each steel component---as-manufactured---will ultimately determine whether each aircraft is safe. As we learned in `c08-structures`, a structure is safe if its applied stress is less than its strength. Therefore, a smaller material strength is a more conservative value for design purposes.
 
 ## Scenarios
 
@@ -49,7 +62,7 @@ In this exercise, we'll focus on the vertical axis: the **source** of variabilit
 <!-- task-end -->
 <!-- solution-begin -->
 - Induced
[…]
