### 2021-08-10 bfe1660
set up model for exercise
