### 2021-08-11 f3b95a2
finish randomization ex
