+++ exercises/e-stat12-randomization-master.Rmd
@@ -5,25 +5,20 @@ date: 2021-08-10
 output: github_document
 time: 30
 reading: 0
+prerequisites:
+  - e-data02-derive
+  - e-stat03-descriptive
 ---
 
-*Purpose*: (Randomization is magic) 
+*Purpose*: You've probably heard that a "control" is important for doing science. If you're lucky, you may have also heard about randomization. These two ideas are the *backbone* of sound data collection. In this exercise, you'll learn the basics about how to plan data collection.
+
+**This is probably __the most important__ lesson in this entire class,** so I hope you do this exercise very carefully!
 
 *Reading*: (None, this is the reading)
 
 ```{r setup}
 library(tidyverse)
-```
 
-# Introduction
-
-# An Example: Fertilizer and Crop Yield
-
-## Scenario
-
-*TODO* (Our expectation in this scenario is that the treatment will have a higher yield than the control)
-
-```{r model-setup}
 ## NOTE: Don't edit this; this sets up the example
 simulate_yield <- function(v) {
   ## Check assertions
@@ -46,6 +41,14 @@ simulate_yield <- function(v) {
 }
 ```
 
+# An Example: Fertilizer and Crop Yield
+
+It's difficult to explain the ideas behind data collection without talking about data to collect, so let's consider a specific example:
+
+Imagine we're testing a fertilizer, and we want to know how much it affects the yield of a specific crop. We have access to a farm, which we section off into six plots. In order to determine the effect the fertilizer has, we need to add fertilizer to some plots, and leave other plots without fertilizer (to serve as a comparison). In scientific jargon, these choices are referred to as the "treatment" and "control". The treatment will have the effect we wish to study, while the control serves as a baseline for a meaningful (quantitative) comparison.
+
+The code below selects a simple arrangement of treatment and control plots.
+
 ```{r define-design}
 ## Define the sequence of treatment (T) and control (C) plots
 experimental_design <- c("T", "T", "T", "C", "C", "C")
@@ -53,12 +56,43 @@ experimental_design <- c("T", "T", "T", "C", "C", "C")
 
 In statistics, the word "design" refers to the "design of *data collection*". The purposeful planning of data collection is called *statistical design of experiments*.
 
-```{r example-run}
+## Visualize the Scenario
+
+The following code visualizes the scenario: how experimental conditions are arranged spatially on our test farm.
+
+```{r vis-design}
+tibble(
+  condition = experimental_design,
+  plot = 1:length(experimental_design)
+) %>% 
+  ggplot(aes(plot, 1)) +
+  geom_label(
+    aes(label = condition, fill = condition),
+    size = 10
+  ) +
+  
+  scale_x_continuous(breaks = 1:6) +
+  scale_y_continuous(breaks = NULL) +
+  theme_minimal() +
+  theme(legend.position = "none") +
+  labs(
+    x = "Plot of Land",
+    y = NULL
+  )
+```
+
+Now let's simulate the results of the experiment!
+
+### __q1__ Simulate the results of this experimental design, and answer the questions under *Observations* below.
+
+```{r q1-task}
+## TODO: Do not edit; run the following code, answer the questions below
 ## For reproducibility, set the seed
 set.seed(101)
 ## Simulate the experimental yields
 experimental_design %>% 
   simulate_yield() %>% 
+  ## Analyze the data
   group_by(condition) %>% 
   summarize(
     yield_mean = mean(yield),
@@ -66,32 +100,76 @@ experimental_design %>%
   )
 ```
 
+*Observations*
 
-## TODO - SCHEMATIC w/ RIVER
+<!-- task-begin -->
+- Which experimental condition do you *expect* to have greater yield? The Treatment `T` or Control `C`?
[…]
