+++ exercises/e-model06-source-master.Rmd
@@ -5,11 +5,12 @@ date: 2021-08-06
 output: pdf_document
 time: 30
 reading: 0
+prerequisites:
+  - e-stat03-descriptive
+  - c08-structures
 ---
 
 *Purpose*: All real data have variability: repeated measurements of "the same" quantity tend to result in different values. To help you recognize different kinds of variability and choose a reasonable analysis procedure based on the kind of variability, you will learn about different *sources* of variability in this exercise.
-
-*Prerequisites*: e-stat03-descriptive, c08-structures
 
 ```{r setup, include=FALSE}
 # knitr options
