### 2021-08-30 feafa5d
add document outline; small edits
