### 2021-10-06 36b026b
add regression test for challenges
