### 2021-10-27 f2e1333
add equality with expectation
