+++ exercises/e-stat02-probability-master.Rmd
@@ -175,9 +175,9 @@ Now that we know what probability is; let's connect the idea to *distributions*.
 
 A continuous *distribution* models probability in terms of an integral
 
-$$\mathbb{P}_{\rho}[X \in A] = \int_{-\infty}^{+\infty} 1_A(x)\rho(x)\,dx$$
+$$\mathbb{P}_{\rho}[X \in A] = \int_{-\infty}^{+\infty} 1_A(x)\rho(x)\,dx = \mathbb{E}_{\rho}[I_A(X)]$$
 
-where $1_A(x)$ is the *indicator function*; a function that takes the value $1$ when $x \in A$, and the value $0$ when $x \not\in A$. The function $\rho(x)$ is the *density* of the random variable---we'll return to this idea in a bit.
+where $1_A(x)$ is the *indicator function*; a function that takes the value $1$ when $x \in A$, and the value $0$ when $x \not\in A$. The $$\mathbb{E}[Y]$$ denotes taking the mean of a random variable; this is also called the *expectation* of a random variable. The function $\rho(x)$ is the *density* of the random variable---we'll return to this idea in a bit.
 
 This definition gives us a geometric way to think about probability; the distribution definition means probability is the area under the density curve within the set $A$.
 
