+++ paper.md
@@ -1,16 +1,17 @@
 ---
-title: 'An Open-Source Active Learning Curriculum for Data Science'
+title: 'An Open-Source Active Learning Curriculum for Data Science in Engineering'
 tags:
   - Active learning
   - Data science
   - R
   - Tidyverse
+  - Engineering
 authors:
   - name: Zachary del Rosario
     orcid: 0000-0003-4676-1692
     affiliation: 1
 affiliations:
- - name: Visiting Professor, Olin College of Engineering
+ - name: Assistant Professor of Engineering and Applied Statistics, Olin College of Engineering
    index: 1
 date:
 bibliography: paper.bib
@@ -28,26 +29,45 @@ encourage students to develop those skills in an open-ended context.
 
 ## Statement of Need
 
-As of writing (Fall 2020), Data Science is an exciting new area of study: Many
+As of writing (Fall 2021), Data Science is an exciting new area of study: Many
 students seek to learn how to obtain, wrangle, and make sense of data, and
 instructors are working to both define and teach this field. An explosion of
 resources is available to both teachers and learners---to name just a
-few---published books [@grolemund2018r], blog posts on Medium [@medium], and
-YouTube channels such as StatQuest [@statquest]. Much of this content is
-excellent and informative.
+few---published books [@grolemund2018r], course materials
+[adhikari2021data8,timbers2021introduction], and YouTube channels such as
+StatQuest [@statquest]. Much of this content is excellent and informative, and
+the broader proliferation of open educational resources contributes positively
+to a more equitable learning environment.
 
 However, one of the most profound findings from the science of learning is the
 importance of *active learning* for student outcomes; *active learning* is a
 pedagogical style that has students actively engaged in the process of learning,
-rather than simply reading a book, skimming a blog post, or watching a video.
-Notably, if teachers held pedagogical design to the same standards as medical
-interventions, a purely passive approach to learning would have been abandoned
-for ethical considerations [@freeman2014active].
+rather than simply reading a book, skimming a blog post, or watching a lecture.
+Notably, active learning has been shown to be superior to a purely passive form
+of instruction [@freeman2014active]. Passive resources are an essential
+component of self-study and are useful for reference, but curricular materials
+for active learning-based instruction are necessary for an instructor seeking to
+maximize student learning. This project seeks to provide active learning
+materials for an intructor seeking to teach data science *to engineers*.
+
+The content in this project has been used to teach a Data Science course at Olin
+College of Engineering; a small engineering college outside Boston that
+emphasizes project-based and active learning pedagogies. Since this course
+caters to engineering students, it has a stronger emphasis on engineering
+problems than is typically found in data science materials. This includes
+analyzing datasets on material properties, basic reliability analysis, and using
+fitted models to make engineering predictions. This curriculum also makes use of
+concepts from statistics education to arm students with the conceptual tools to
+recognize and handle different forms of variability that typically arise in
+engineering problems [@wild1999statistical,@aggarwal2021qualitative].
 
 This work provides a two-tiered set of content for active learning of data
 science. This content is appropriate for both instructors seeking to teach data
 science using a flipped classroom model [@bishop2013flipped], and for individual
-learners seeking a self-study course in data science.
+learners seeking a self-study course in data science. Additionally, since the
+materials herein are released under a permissive open-source license, other
+instructors are welcome (and encouraged) to rearrange and modify the content
+provided.
 
 ## Pedagogical Design
 
