+++ paper.bib
@@ -158,3 +158,43 @@ gplot2'},
   note = {R package version 0.1.8},
   url = {https://CRAN.R-project.org/package=modelr},
 }
+
+@inproceedings{aggarwal2021qualitative,
+  title={A Qualitative Study of Engineering Students' Reasoning About Statistical Variability},
+  author={Aggarwal, Riya and Flynn, Mira and Daitzman, Sam and Lam, Diane and del Rosario, Zachary Riggins},
+  booktitle={2021 Fall ASEE Middle Atlantic Section Meeting},
+  year={2021}
+}
+
+@book{adhikari2021data8,
+  title={Computational and Inferential Thinking: The Foundations of Data Science},
+  author={Adhikari, Ani and DeNero, John and Wagner, David},
+  year={2021},
+  url={https://inferentialthinking.com/chapters/intro.html},
+}
+
+@book{timbers2021introduction,
+  title={Data Science: A First Introduction},
+  author={Timbers, Tiffany-Anne and Campbell, Trevor and Lee, Melissa},
+  year={2021},
+  url={https://ubc-dsci.github.io/introduction-to-datascience/},
+}
+
+@manual{adenbuie2021learnr,
+  title = {learnr: Interactive Tutorials for R},
+  author = {Garrick Aden-Buie and Barret Schloerke and JJ Allaire},
+  year = {2021},
+  url = {https://rstudio.github.io/learnr/},
+  note = {https://rstudio.github.io/learnr/, https://github.com/rstudio/learnr},
+}
+
+@article{wild1999statistical,
+  title={Statistical thinking in empirical enquiry},
+  author={Wild, Chris J and Pfannkuch, Maxine},
+  journal={International statistical review},
+  volume={67},
+  number={3},
+  pages={223--248},
+  year={1999},
+  publisher={Wiley Online Library}
+}
