### 2021-12-14 zdelrosario/data-science-curriculum#25 @zdelrosario
@pschloss good point about articulating a specific need; I have updated the paper to emphasize that the present content has a stronger emphasis on engineering content than is typical in a data science curriculum. Is is appropriate (and possible) to modify the paper title to reflect this?

I also agree that my prior use of the Freeman et al. reference was unproductive; I have rephrased.

@firasm I have added some of the references you suggested; highlighted the potential of passive resources for self-study (and reference), and added your point about the equity benefits of OER.
