### 2021-12-15 c7a611c
add infrastructure for book
