### 2021-12-15 c1a4656
various fixes for proper execution
