+++ book/_bookdown.yml
@@ -1,0 +1,1 @@
+new_session: yes
