+++ exercises/e-rep04-directories-master.Rmd
@@ -7,6 +7,8 @@ time: 10
 reading: 10
 ---
 
+# Reproducibility: Structuring Directories
+
 *Purpose*: A key part of being successful in data science is *getting
 organized*. In this exercise, you'll set yourself up for future success by
 learning how to *structure* a project in terms of its folder hierarchy: its
@@ -15,7 +17,7 @@ subdirectories, and protect your raw data.
 
 *Reading*: [Project Structure](https://reproducible-science-curriculum.github.io/organization-RR-Jupyter/02-directories/) (Skim: Provided so you can see an example of a well-structured project directory)
 
-# Reading: Structuring a Project
+## Reading: Structuring a Project
 <!-- -------------------------------------------------- -->
 
 In this exercise, we'll assume that you are working on a project.
@@ -68,23 +70,23 @@ writes out a processed version of the data. You can then load the processed data
 in later scripts, and all of your processing steps will be permanently
 documented.
 
-# Practice: Set up a project
+## Practice: Set up a project
 <!-- -------------------------------------------------- -->
 
 Let's set you up for future success by creating a directory for your Project 1.
 
-__q1__ Create a project directory in your personal data science rep called
+### __q1__ Create a project directory in your personal data science rep called
 `p01-name`. If you already know what you're going to work on, feel free to
 replace `name` with something sensible. If you don't know what to call it yet,
 don't worry---you can change it later!
 
-__q2__ Create a `data` directory under `p01-name`; you should then have
+### __q2__ Create a `data` directory under `p01-name`; you should then have
 `p01-name/data`. This is where you will put any data files you read or write in the project.
 
 There's a *trick* to committing an empty folder with Git. We'll need to
 introduce some kind of file to preserve an empty directory structure.
 
-__q3__ Create an empty file called `.gitignore` under your `data` folder. You
+### __q3__ Create an empty file called `.gitignore` under your `data` folder. You
 can do this from the root of your data science repo by using the following from
 your command line:
 

[…]
