### 2021-12-16 c98ce4b
rm extra yaml for bookdown
