### 2021-12-16 3a185aa
don't evaluate auth call in solutions
