+++ challenges/c04-gapminder-assignment.Rmd
@@ -1,0 +1,156 @@
+---
+title: "Gapminder"
+author: "(Your name here)"
+date: 2020-
+output:
+  github_document:
+    toc: true
+---
+
+*Purpose*: Learning to do EDA well takes practice! In this challenge you'll further practice EDA by first completing a guided exploration, then by conducting your own investigation. This challenge will also give you a chance to use the wide variety of visual tools we've been learning.
+
+<!-- include-rubric -->
+# Grading Rubric
+<!-- -------------------------------------------------- -->
+
+Unlike exercises, **challenges will be graded**. The following rubrics define how you will be graded, both on an individual and team basis.
+
+## Individual
+<!-- ------------------------- -->
+
+| Category | Needs Improvement | Satisfactory |
+|----------|----------------|--------------|
+| Effort | Some task __q__'s left unattempted | All task __q__'s attempted |
+| Observed | Did not document observations, or observations incorrect | Documented correct observations based on analysis |
+| Supported | Some observations not clearly supported by analysis | All observations clearly supported by analysis (table, graph, etc.) |
+| Assessed | Observations include claims not supported by the data, or reflect a level of certainty not warranted by the data | Observations are appropriately qualified by the quality & relevance of the data and (in)conclusiveness of the support |
+| Code Styled | Violations of the [style guide](https://style.tidyverse.org/) hinder readability | Code sufficiently close to the [style guide](https://style.tidyverse.org/) |
+
+## Team
+<!-- ------------------------- -->
+
+| Category | Needs Improvement | Satisfactory |
+|----------|----------------|--------------|
+| Documented | No team contributions to Wiki | Team contributed observations to Wiki |
+| Referenced | No team references in Wiki | At least one reference in Wiki to member report(s) |
+| Relevant | References unrelated to observations, or difficult to find related analysis based on reference text | Reference text clearly points to relevant analysis with active hyperlinks |
+
+## Due Date
+<!-- ------------------------- -->
+
+All the deliverables stated in the rubrics above are due **at midnight** before the day of the class discussion of the challenge. See the [Syllabus](https://docs.google.com/document/d/1jJTh2DH8nVJd2eyMMoyNGroReo0BKcJrz1eONi3rPSc/edit?usp=sharing) for more information.
+
+```{r setup}
+library(tidyverse)
+library(gapminder)
+
+```
+
+*Background*: [Gapminder](https://www.gapminder.org/about-gapminder/) is an
+independent organization that seeks to educate people about the state of the
+world. They seek to counteract the worldview constructed by a hype-driven media
+cycle, and promote a "fact-based worldview" by focusing on data. The dataset
+we'll study in this challenge is from Gapminder.
+
+# Guided EDA
+<!-- -------------------------------------------------- -->
+
+First, we'll go through a round of *guided EDA*. Try to pay attention to the high-level process we're going through---after this guided round you'll be responsible for doing another cycle of EDA on your own!
+
+__q0__ Perform your "first checks" on the dataset. What variables are in this
+dataset?
+
+```{r q0-task}
+## TASK: Do your "first checks" here!
+```
+
+**Observations**:
+
+- Write the variable names here
+
+__q1__ Determine the most and least recent years in the `gapminder` dataset.
+
+*Hint*: You may need to use the `pull()` function to get a vector out of a tibble.
+
+```{r q1-task}
+## TASK: Find the largest and smallest values of `year` in `gapminder`
+year_max <- NA_real_
+year_min <- NA_real_
+```
+
+Use the following test to check your work.
+
+```{r q1-tests}
+## NOTE: No need to change this
[…]
