+++ d45-e-model04-logistic-solution.md
@@ -412,7 +412,7 @@ df_basic %>%
 ## # A tibble: 1 x 4
 ##      TP    FP    TN    FN
 ##   <int> <int> <int> <int>
-## 1    37    29    25     6
+## 1    38    31    19     9
 ```
 
 These numbers don't mean a whole lot on their own; we'll use them to compare performance across models. Next you'll practice using R functions to carry out logistic regression for classification, and build a model to compare against this basic one.
@@ -535,7 +535,7 @@ df_q4 %>%
 ## # A tibble: 1 x 4
 ##      TP    FP    TN    FN
 ##   <int> <int> <int> <int>
-## 1    29     6    48    14
+## 1    34     5    45    13
 ```
 
 **Observations**:
