+++ d16-e-stat02-probability-solution.md
@@ -87,7 +87,7 @@ df_z %>%
 ## # A tibble: 1 x 3
 ##   count_total count_A    fr
 ##         <int>   <int> <dbl>
-## 1         100      97  0.97
+## 1         100      93  0.93
 ```
 
 Now it's your turn!
@@ -108,7 +108,7 @@ df_z %>%
 ## # A tibble: 1 x 3
 ##   count_total count_A    fr
 ##         <int>   <int> <dbl>
-## 1         100      51  0.51
+## 1         100      47  0.47
 ```
 
 **Observations**:
@@ -160,9 +160,9 @@ map_dfr(
 ##   count_total count_A    fr
 ##         <int>   <int> <dbl>
 ## 1          10       4 0.4  
-## 2         100      51 0.51 
-## 3        1000     527 0.527
-## 4       10000    4958 0.496
+## 2         100      50 0.5  
+## 3        1000     493 0.493
+## 4       10000    4976 0.498
 ```
 
 This is because *probability* is actually defined[1] in terms of the limit
@@ -193,10 +193,10 @@ map_dfr(
 ## # A tibble: 4 x 3
 ##   count_total count_A    fr
 ##         <int>   <int> <dbl>
-## 1          10      10 1    
-## 2         100      92 0.92 
-## 3        1000     950 0.95 
-## 4       10000    9503 0.950
+## 1          10       9 0.9  
+## 2         100      95 0.95 
+## 3        1000     954 0.954
+## 4       10000    9509 0.951
 ```
 
 **Observations**:
