+++ d12-e-stat01-densities-solution.md
@@ -253,7 +253,7 @@ rnorm(5)
 ```
 
 ```
-## [1] -0.6638926  1.2545052  0.3744894 -0.6293435  1.4608634
+## [1]  0.75248272  0.55276910  0.61945397 -0.42165563 -0.01060184
 ```
 
 What this means is we'll get a slightly different picture every time we draw a sample. **This is the challenge with randomness**: Since we could have drawn a different set of samples, we need to know the degree to which we can trust conclusions drawn from data. Being *statistically literate* means knowing how much to trust your data.
