### 2021-12-18 78cb567
ensure build push of exercise Rmd files
