+++ exercises_sequenced/d12-e-rep04-directories-assignment.Rmd
@@ -1,0 +1,124 @@
+---
+title: "Reproducibility: Structuring Directories"
+author: Zachary del Rosario
+date: 2020-07-23
+output: github_document
+time: 10
+reading: 10
+---
+
+# Reproducibility: Structuring Directories
+
+*Purpose*: A key part of being successful in data science is *getting
+organized*. In this exercise, you'll set yourself up for future success by
+learning how to *structure* a project in terms of its folder hierarchy: its
+*directories*. You'll learn how to isolate your projects, structure your
+subdirectories, and protect your raw data.
+
+*Reading*: [Project Structure](https://reproducible-science-curriculum.github.io/organization-RR-Jupyter/02-directories/) (Skim: Provided so you can see an example of a well-structured project directory)
+
+## Reading: Structuring a Project
+<!-- -------------------------------------------------- -->
+
+In this exercise, we'll assume that you are working on a project.
+
+## Isolate your project
+
+While there are [many
+ways](https://medium.com/javascript-in-plain-english/how-to-organize-project-files-4b9d9fd02e73)
+to *organize* a given project, you should **definitely** organize your work into
+individual *projects*. Each project should have its own directory, and should
+generally have all its related files under that directory.
+
+For instance, `data-science-curriculum` is in my mind an entire project, containing both exercises and challenges. Thus for me, it makes sense to organize all of these *subdirectories* under `data-science-curriculum`.
+
+## Structure your subdirectories
+
+Subdirectories are folders (directories!) *under* the main (root) project
+folder. For instance, the `data-science-curriculum` has subdirectories
+`exercises`, `challenges`, and more.
+
+Each subdirectory should have some *purpose*. For instance, there are both
+`exercises` and `exercises_sequenced` directories. I build the exercises in
+`exercises` out-of-sequence, because I don't want to have to decide on which
+specific day each exercise should occur *before* I start building it. However,
+it's useful for you the student to be able to see all the exercises in daily
+order, so I have an additional directory where I can place the sequenced
+exercises.
+
+Knitr will *automatically* create some sensible subdirectories for report files;
+for instance, when you knit a document knitr will automatically create a
+`filename_files` directory with image files. This helps prevent clutter in your
+directory, and places the file adjacent to the source file in your
+(sub)directory.
+
+Your project can have deeper levels of *hierarchy*: *subdirectors under
+subdirectories*. For instance, the `exercises` folder has a subdirectory
+`exercises/data`; this is where exercise datasets live. In your own projects,
+you should keep your data *unaltered* in some kind of data directory.
+
+## Protect your data!
+
+**Never** make alterations to your raw data! Even if your data have errors, it's
+far better to *document* those errors somewhere, such that you have a papertrail
+for what changed and why. It is far better practice to keep your data unedited,
+and simply write *processed* data to an additional file.
+
+If your data needs preprocessing---say to fix errors or simply to wrangle
+data---write a preprocessing Rmarkdown file that takes in the raw data, and
+writes out a processed version of the data. You can then load the processed data
+in later scripts, and all of your processing steps will be permanently
+documented.
+
+## Practice: Set up a project
+<!-- -------------------------------------------------- -->
+
+Let's set you up for future success by creating a directory for your Project 1.
+
+### __q1__ Create a project directory in your personal data science rep called
+`p01-name`. If you already know what you're going to work on, feel free to
+replace `name` with something sensible. If you don't know what to call it yet,
+don't worry---you can change it later!
+
[…]
