+++ exercises_sequenced/d28-e-stat06-clt-assignment.Rmd
@@ -1,0 +1,327 @@
+---
+title: "Stats: The Central Limit Theorem and Confidence Intervals"
+author: Zach del Rosario
+date: 2020-06-23
+output: github_document
+time: 45
+reading: 0
+---
+
+# Stats: The Central Limit Theorem and Confidence Intervals
+
+*Purpose*: When studying sampled data, we need a principled way to report our results with their uncertainties. Confidence intervals (CI) are an excellent way to summarize results, and the central limit theorem (CLT) helps us to construct these intervals.
+
+*Reading*: (None, this is the reading)
+
+*Topics*: The central limit theorem (CLT), confidence intervals
+
+```{r setup}
+library(tidyverse)
+library(nycflights13)
+```
+
+## Central Limit Theorem
+<!-- -------------------------------------------------- -->
+
+Let's return to a result from `e-stat04-population`:
+
+```{r recall}
+## NOTE: No need to edit this
+set.seed(101)
+n_observations <- 9
+n_samples <- 5e3
+
+df_samp_unif <-
+  map_dfr(
+    1:n_samples,
+    function(id) {
+      tibble(
+        Z = runif(n_observations),
+        id = id
+      )
+    }
+  )
+
+df_samp_unif %>%
+  group_by(id) %>%
+  summarize(stat = mean(Z)) %>%
+
+  ggplot(aes(stat)) +
+  geom_histogram() +
+  labs(
+    x = "Estimated Mean",
+    title = "Sampling Distribution: Estimated Mean",
+    caption = "Population: Uniform"
+  )
+```
+
+If you said that the sampling distribution from the exercise above looks roughly normal, then you are correct! This is an example of the [central limit theorem](https://en.wikipedia.org/wiki/Central_limit_theorem), a central idea in statistics. Here we'll introduce the central limit theorem (CLT), use it to approximate the sampling distribution for the sample mean, and in turn use that to construct an approximate *confidence interval*.
+
+For populations satisfying mild conditions[1], the sample mean $\overline{X}$ converges to a normal distribution as the sample size $n$ approaches infinity. Specifically
+
+$$\overline{X} \stackrel{d}{\to} N(\mu, \sigma^2 / n),$$
+
+where $\mu$ is the mean of the population, $\sigma$ is the standard deviation of the population, and $\stackrel{d}{\to}$ means [*converges in distribution*](https://en.wikipedia.org/wiki/Convergence_of_random_variables#Convergence_in_distribution), a technical definition that is beyond the scope of this lesson.
+
+Below I simulate sampling from a uniform distribution and compute the mean at different sample sizes to illustrate the CLT:
+
+```{r gen-clt}
+## NOTE: No need to change this!
+set.seed(101)
+n_repl <- 5e3
+
+df_clt <-
+  map_dfr(
+    1:n_repl,
+    function(id) {
+      map_dfr(
+        c(1, 2, 9, 81, 729),
+        function(n) {
+          tibble(
+            Z = runif(n),
+            n = n,
+            id = id
+          )
+        }
+      )
+    }
+  ) %>%
+  group_by(n, id) %>%
+  summarize(mean = mean(Z), sd = sd(Z))
+```
+
+Let's visualize the sampling distribution for each sample size:
+
+```{r vis-clt}
+df_clt %>%
+  ggplot(aes(mean)) +
+  geom_density() +
+  facet_wrap(~n, scales = "free")
+```
+
+At just `1` our sample mean is $X_1 / 1$---we're just drawing single observations from the population, so we see a uniform. At `2` we something that looks like a tent. By `9` samples we see a distribution that looks quite normal.
+
[…]
