+++ exercises_sequenced/d01-e-rep02-create-assignment.Rmd
@@ -1,0 +1,39 @@
+---
+title: "Reproducibility: Creating a GitHub Repository"
+author: Zachary del Rosario
+date: 2020-06-24
+output: github_document
+time: 5
+reading: 10
+---
+
+# Reproducibility: Creating a GitHub Repository
+
+*Purpose*: In this class you'll turn in all your work using GitHub. I created a repo for you at the beginning of class, but you should know how to make your own, e.g. for future projects you might do.
+
+*Reading*: Follow all the steps in [Creating a Repository](https://swcarpentry.github.io/git-novice/03-create/index.html). *Hint*: You can `Ctrl + Click` (`CMD + Click`) to follow links in RStudio.
+
+*Note*: For the steps in the reading, I recommend using the Terminal in RStudio. This should help ensure you have access to Git.
+
+*Topics*: Creating a GitHub repository
+
+## Set Up Your Repository
+<!-- -------------------------------------------------- -->
+
+- Log into [GitHub](https://github.com/) and click the "New" button to start creating a new repository.
+
+![New](./images/rep02-new.png)
+
+- Give your a repository a name (such as `data-science-work`), enable "Initialize this repository with a README", then click "Create repository". This will take you to your new repository.
+
+*Important Note*: Make sure to click "Initialize this repository with a README"! This will help when you go to clone your repository later!
+
+![Info](./images/rep02-info-readme.png)
+
+Next, follow the steps in [e-rep01-intro-git](https://github.com/zdelrosario/data-science-curriculum/blob/master/exercises/e-rep01-intro-git.md) to clone your new repository (make sure to target your new repository, though!).
+
+<!-- include-exit-ticket -->
+# Exit Ticket
+<!-- -------------------------------------------------- -->
+
+Once you have completed this exercise, make sure to fill out the **exit ticket survey**, [linked here](https://docs.google.com/forms/d/e/1FAIpQLSeuq2LFIwWcm05e8-JU84A3irdEL7JkXhMq5Xtoalib36LFHw/viewform?usp=pp_url&entry.693978880=e-rep02-create-assignment.Rmd).

[…]
