### 2021-12-18 zdelrosario/data-science-curriculum#26 @zdelrosario
@pschloss and @firasm this took *a while* to figure out, but I believe I've implemented a solution: I've set up one GitHub actions to auto-build a [book](https://zdelrosario.github.io/data-science-curriculum/index.html) to serve as an easily-viewed solution manual, and another action to auto-run the make utilities and create a derived [build branch](https://github.com/zdelrosario/data-science-curriculum/tree/build). This obviates the need for a user to run make locally, making the content far more accessible to a novice user. I've updated the [readme](https://github.com/zdelrosario/data-science-curriculum/blob/build/README.md) to match the new instructions; a user now need only download and unzip an archive.
