### 2021-12-18 13698d7
add more details to book landing page
