+++ book/index.Rmd
@@ -5,9 +5,21 @@ site: bookdown::bookdown_site
 documentclass: book
 output:
   bookdown::gitbook: default
-  #bookdown::pdf_book: default
 ---
 
 # Data Science: Exercise Solutions
 
-(This is a collection of the solution material.)
+These are the exercise solutions for this [data science curriculum](https://github.com/zdelrosario/data-science-curriculum). Note that you should *attempt to do* each exercise before consulting the solutions! However, if you get stuck, you should make sure to consult these materials.
+
+## How to Set Up These Materials
+
+1. (Setup) Complete [this exercise](https://zdelrosario.github.io/data-science-curriculum/setup-rstudio.html) to install RStudio.
+2. (Setup) Download and unzip [this archive](https://github.com/zdelrosario/data-science-curriculum/archive/refs/heads/build.zip) to obtain the curriculum materials.
+3. (Setup) Open the folder you unzipped as a project in RStudio (`File > Open Project...`).
+4. (Learn) Work through the exercise files in the `exercises_sequenced/` folder at your own pace to learn data science skills.
+5. (Learn) Use the challenges in the `challenges/` folder to put your new skills to use.
+
+## How to Use This Guide
+
+- Download and install the materials, as described in the steps above.
+- Work through the exercises; if you get stuck, take a look at this solution guide to get unstuck.
