+++ d01-e-setup02-functions-solution.md
@@ -28,7 +28,7 @@ rnorm
 ```
 ## function (n, mean = 0, sd = 1) 
 ## .Call(C_rnorm, n, mean, sd)
-## <bytecode: 0x7f9ec667c9d0>
+## <bytecode: 0x7fc6f5cc05d0>
 ## <environment: namespace:stats>
 ```
 
