### 2021-12-21 e603fd7
update curriculum to match exercises
