+++ curriculum.md
@@ -4,23 +4,21 @@
 
 By the end of this course, students will be able to...
 
-## (Setup)
+## Setup
 
-Core
 - install Rstudio and other data science packages (e-setup00-install)
 - load packages to access useful functions (e-setup01-packages)
 - access documentation to use R functions (e-setup02-functions)
 - access vignettes to learn new functions (e-setup03-docs)
 - use Rstudio shortcuts to improve productivity (e-setup04-rstudio-shortcuts)
 
-## (Data)
+## Data
 
-Core
-- Recognize
-  - basic data objects: vectors, types (e-setup05-vectors, e-setup06-types)
 - Isolate
   - data using `filter` (e-data01-isolate)
   - variables using `select` (e-data01-isolate)
+- Recognize
+  - basic data objects: vectors, types (e-setup05-vectors, e-setup06-types)
 - Derive
   - new variables with `mutate` (e-data02-derive)
   - summaries with `group_by` and `summarize` (e-data02-derive)
@@ -32,26 +30,24 @@ Core
   - by `join` of multiple tables (e-data05-join)
 - Load
   - data from a flat file, e.g. a comma-separated value (csv) file (e-data00-basics)
+  - data from an Excel sheet (e-data09-readxl)
   - data from googlesheets (e-data12-simple-pipeline)
 - Wrangle
   - data from a messy source (e-data09-readxl, e-data13-cleaning)
+  - strings with stringr (e-data06-strings)
+  - complex code into readable, functional pipelines (e-data07-placeholders)
   - loops into functional calls using purrr (e-data10-map)
   - factors with forcats (e-data11-factors)
 - Liberate
-  - data from images with WebPlotDigitizer
-  - data from printed tables with Tabula
+  - data from images with WebPlotDigitizer (e-data14-webplotdigitizer)
+  - data from printed tables with Tabula (e-data15-tabula)
 
-Bonus
-- Specialty
-  - (lubridate)
+## Visualization
 
-## (Visualization)
-
-Core
 - State and use the basic components of the *grammar of graphics*:
-  - geometry (many)
+  - geometry (many!)
   - aesthetics (e-vis00-basics)
-  - theme (e-vis07-themes)
+  - theme (e-vis06-themes)
   - layers (e-vis04-scatterplot)
 - Create
   - bar charts (e-vis01-bar-charts)
@@ -60,81 +56,62 @@ Core
   - scatterplots (e-vis04-scatterplot)
   - line plots (e-vis05-lines)
   - small multiple plots (e-vis06-multiples)
-  - tweaks to visual aesthetics (e-vis07-themes)
+  - tweaks to visual aesthetics (e-vis06-themes)
 - Design
-  - EDA-quality graphics (many)
-  - presentation-ready graphics (e-vis07-themes, e-comm01-story-basics)
+  - EDA-quality graphics (many!)
+  - presentation-ready graphics (e-vis06-themes, e-comm01-story-basics)
+  - comparisons using small multiples (e-vis08-multiples)
+  - improved visuals through an iterative approach (e-vis09-improve)
 - Criticize
-  - data visualizations based on *the visual hierarchy*
-  - data visualizations based on their intended audience
+  - data visualizations based on *the visual hierarchy* (e-vis07-perception)
 
-Bonus
-- (Geospatial with sf)
-- (Shiny apps)
-- (Leaflet maps)
+## Statistical Literacy
 
-## (Statistical Literacy)
-
-Core
 - Distinguish
   - between error and uncertainty (c02-michelson, e-stat07-error-bias)
-  - between variability and uncertainty
   - between populations and samples (e-stat04-population)
 - Describe
   - a random quantity with a distribution (e-stat01-distributions)
[…]
