+++ paper.bib
@@ -198,3 +198,22 @@ gplot2'},
   year={1999},
   publisher={Wiley Online Library}
 }
+
+@book{tukey1977eda,
+  title     = {Exploratory data analysis},
+  author    = {Tukey, John W.},
+  volume    = {2},
+  year      = {1977},
+  publisher = {Reading, Mass.}
+}
+
+@article{moore1997new,
+  title={New pedagogy and new content: The case of statistics},
+  author={Moore, David S},
+  journal={International statistical review},
+  volume={65},
+  number={2},
+  pages={123--137},
+  year={1997},
+  publisher={Wiley Online Library}
+}
