### 2021-12-21 76ff7ae
add instructions for editing source files
