### 2021-12-21 #117 @zdelrosario
(Re-building paper with suggested edits.)
