### 2021-12-21 e3466e7
fix inline citations
