+++ paper.bib
@@ -161,7 +161,7 @@ gplot2'},
 
 @inproceedings{aggarwal2021qualitative,
   title={A Qualitative Study of Engineering Students' Reasoning About Statistical Variability},
-  author={Aggarwal, Riya and Flynn, Mira and Daitzman, Sam and Lam, Diane and del Rosario, Zachary Riggins},
+  author={Aggarwal, Riya and Flynn, Mira and Daitzman, Sam and Lam, Diane and {del Rosario}, Zachary Riggins},
   booktitle={2021 Fall ASEE Middle Atlantic Section Meeting},
   year={2021}
 }
