+++ d01-e-setup02-functions-solution.md
@@ -28,7 +28,7 @@ rnorm
 ```
 ## function (n, mean = 0, sd = 1) 
 ## .Call(C_rnorm, n, mean, sd)
-## <bytecode: 0x7fb90c0ea7d0>
+## <bytecode: 0x7fd5ddeef9d0>
 ## <environment: namespace:stats>
 ```
 
