### 2021-12-21 f1425dd
rm outdated note about core vs bonus
