### 2021-12-21 #117 @zdelrosario
(Re-building paper to correct typos; collect small edits)
