### 2021-12-21 abdc480
fix multi-ref citation syntax
