### 2021-12-21 #117 @zdelrosario
@ttimbers , @pschloss , @firasm

Many thanks to @pschloss and @firasm for the detailed and helpful comments. *Clearly* it has taken me a long while to respond to this review. A major obstacle for me was finding the time to teach myself some of the GitHub Action trickery necessary to fully address the reviewer comments. However, I believe I have found a solution that substantially improves the infrastructure of the repository, which will make these materials much easier for both teachers and learners to use the content. Please see links to reviewer comments below, and my responses to those comments:

Comments from @pschloss

- The [License](https://github.com/zdelrosario/data-science-curriculum/issues/24)
  - As suggested, I have changed to a [CC-SA](https://github.com/zdelrosario/data-science-curriculum/blob/master/LICENSE) license
- The [Statement of Need](https://github.com/zdelrosario/data-science-curriculum/issues/25)
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) As suggested, I have removed the unproductive dunk on passive learning....
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) I have more-clearly articulated the need/niche that these materials fill: This is an active-learning data science curriculum aimed at engineering students
  - **With that in mind, is is possible to change the title of the paper submission?** I think that 'An Open-Source Active Learning Curriculum for Data Science in Engineering' captures the more-focused need.
- The [Installation Instructions](https://github.com/zdelrosario/data-science-curriculum/issues/26)
  - I have made substantial changes to the infrastructure and installation instructions; the materials now auto-update to a user-friendly build via GitHub Action. This is reflected in the updated ['How to Use This Repo'](https://github.com/zdelrosario/data-science-curriculum#how-to-use-this-repo) instructions in the root [readme](https://github.com/zdelrosario/data-science-curriculum).
- [Usage](https://github.com/zdelrosario/data-science-curriculum/issues/27)
  - Related to above; the [instructions](https://github.com/zdelrosario/data-science-curriculum#how-to-use-this-repo) now point to a single archive from the (new) `build` branch, which drastically simplifies the process of setting up the materials.
  - As-suggested, I have also added a `.Rproj` [file](https://github.com/zdelrosario/data-science-curriculum/blob/master/data-science-curriculum.Rproj) to the repo (and build). This is no longer necessary to find the correct directory, but it does help with quality-of-life issues (to my knowledge, the Git interface for Rstudio only works when a project is open).

Comments from @firasm

- [Statement of need](https://github.com/zdelrosario/data-science-curriculum/issues/25)
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) As suggested, I made my references to existing materials a bit more selective
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) Made the suggested edit(s) around the Freeman reference
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) Made a note about OER
- [Installation Instructions](https://github.com/zdelrosario/data-science-curriculum/issues/26)
  - Made substantial improvement to these instructions: Automated a [`build` branch](https://github.com/zdelrosario/data-science-curriculum/tree/build) through GitHub Actions to produce a user-friendly zip archive. This obviates the need for a local run of `make`, and automatically down-selects to just the `challenges/` and `exercises_sequenced/` directories.
    - These changes are reflected in the [use instructions](https://github.com/zdelrosario/data-science-curriculum#how-to-use-this-repo).
  - I have also built a [web version](https://zdelrosario.github.io/data-science-curriculum/index.html) of the solutions; this is a more handy and accessible version of the solutions. Also, as you suggested, this is a nice way to advertise the content!
- [Usage](https://github.com/zdelrosario/data-science-curriculum/issues/27)
  - The need to run `make` is obviated by the updates to Installation Instructions mentioned above.
  - The confusing discrepancies between `exercises/` and `exercises_sequenced/` is obviated (for learners) by the updates to Installation Instructions mentioned above.
  - (In the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md)) I also describe how an instructor can adapt these materials.
- [New Release](https://github.com/zdelrosario/data-science-curriculum/issues/28)
  - As suggested, I have cut a [new release](https://github.com/zdelrosario/data-science-curriculum/releases/tag/v0.2).
- [Pedagogical Design](https://github.com/zdelrosario/data-science-curriculum/issues/29)
  - Added a direct link to the learning outcomes in the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md).
  - Added the suggested citation for flipped classrooms.
- [Sequence](https://github.com/zdelrosario/data-science-curriculum/issues/31)
  - I believe this confusion is obviated (for learners) by the updates to Installation Instructions mentioned above.
- [Overall Feedback](https://github.com/zdelrosario/data-science-curriculum/issues/30)
  - As @pschloss suggested I changed to a CC-SA license; I hope this preserves the same level of openness as the MIT license...!
  - I have also built a [web version](https://zdelrosario.github.io/data-science-curriculum/index.html) of the solutions; as you suggested, this is a nice way to advertise the content!
  - Added further description of the materials and sequence in the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md).
  - Added a description of how these materials have been used---and how one might adapt them---in the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md).
  - Added the "development story" in the [paper](https://github.com/zdelrosario/data-science-curriculum/blob/master/paper.md).
  - Key models are listed in the [learning outcomes](https://github.com/zdelrosario/data-science-curriculum/blob/master/curriculum.md) document (now linked from the paper). Note that this also points to specific exercises that implement each learning outcome.
  - The [learning outcomes](https://github.com/zdelrosario/data-science-curriculum/blob/master/curriculum.md) is now linked from the paper; this provides an overview of the course content without needing to reference every individual file.
