### 2022-01-08 #117 @firasm
This is fantastic @zdelrosario - well done. It's been a long time but I think you've addressed all my comments, so I'm happy!

The content is much more discoverable now. I except it will also make it easy for others to use your work as well. I'm not sure what your overall goal is with the project, but if you like, you can convert the repo to a ["template" repo](https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-template-repository) (this is different from forking) so others can get an exact duplicate of your repo with all the actions and the machinery.
