### 2022-01-23 #117 @zdelrosario
@firasm Thanks! I'll give the template repository option some thought: Helping others use the action infrastructure definitely sounds useful, but doesn't that also prevent forking? I would also like to allow folks to easily contribute to improving the materials, so it seems like a tradeoff to me.

@pschloss Thanks!

@ttimbers Please let me know if there is anything else I need to do to finish up this revision.
