### 2022-01-23 #117 @firasm
Template repos can still be forked - they are mostly independent.
