### 2022-01-24 zdelrosario/data-science-curriculum#37 @zdelrosario
Add documentation on use cases
Should document how to use this repo for different use cases; for instance, when to use the template (setting up with GitHub action infrastructure) and when to fork (to make a PR).
