### 2022-01-24 #117 @zdelrosario
@firasm good to know! No reason not to make it a template repo then. I'll make sure to [add some documentation](https://github.com/zdelrosario/data-science-curriculum/issues/37) on this use-case.
