### 2022-02-28 #117 @zdelrosario
@ttimbers sure thing! I've already created a [release](https://github.com/zdelrosario/data-science-curriculum/releases/tag/v0.2) on GitHub, and just uploaded an archived copy of this to [Zenodo](https://zenodo.org/record/6314141#.Yh1Rud9OkWo). This has minted the doi: 10.5281/zenodo.6314141

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6314141.svg)](https://doi.org/10.5281/zenodo.6314141)
