+++ paper.bib
@@ -42,8 +42,7 @@
   volume={30},
   number={9},
   pages={1--18},
-  year={2013},
-  doi={10.18260/1-2--22585}
+  year={2013}
 }
 
 @book{lang2016small,
