+++ paper.bib
@@ -195,7 +195,8 @@ gplot2'},
   number={3},
   pages={223--248},
   year={1999},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2307/1403699},
 }
 
 @book{tukey1977eda,
@@ -214,5 +215,6 @@ gplot2'},
   number={2},
   pages={123--137},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2307/1403333}
 }
