### 2022-03-09 #117 @zdelrosario
Hmm, I checked that last invalid DOI, and it appears to be valid in the raw [bibtex file](https://github.com/zdelrosario/data-science-curriculum/commit/cd2869fa04c401425455abc01db496fd1852402b), reproduced below.

```
  doi={10.18260/1-2--22585}
```

Looks like that double dash `--` might be getting converted to a single dash?

Also couldn't find a DOI for Tukey's EDA, but that's a rather old reference, so I'm not surprised.
