### 2022-03-12 #117 @ttimbers
@zdelrosario - apologies for the delay! Was consulting with more experienced folks on what to do here. `10.18260/1-2--22585` looks like a valid DOI and so we can ignore the check saying that it is missing (the checks are not perfect). And yes, as Tukey's book is from quite a while ago, it has no DOI. So again, we can ignore the missing check on that.
