### 2022-03-12 b96e7d6
restore doi for Bishop et al.
