### 2022-03-15 #117 @zdelrosario
Thanks @labarba , @ttimbers , @firasm , @pschloss , and everyone at the JOSE team! This whole review process has *significantly* improved the teaching materials, and was overall a very positive experience!
